requirements.txt
streamlit==1.29.0
pyresparser==1.0.6
pdfminer.six==20221105
mysql-connector-python==8.2.0
nltk==3.8.1
pandas==2.1.3
numpy==1.25.2
pillow==10.0.1
plotly==5.17.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2

